switch the otelhttp semconv tests to use metricdatatest
