--- instrumentation/net/http/otelhttp/internal/semconv/common_test.go
@@ -1,155 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package semconv
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"strconv"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
-)
-
-type testServerReq struct {
-	hostname   string
-	serverPort int
-	peerAddr   string
-	peerPort   int
-	clientIP   string
-}
-
-func testTraceRequest(t *testing.T, serv HTTPServer, want func(testServerReq) []attribute.KeyValue) {
-	t.Helper()
-
-	got := make(chan *http.Request, 1)
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		got <- r
-		close(got)
-		w.WriteHeader(http.StatusOK)
-	}
-
-	srv := httptest.NewServer(http.HandlerFunc(handler))
-	defer srv.Close()
-
-	srvURL, err := url.Parse(srv.URL)
-	require.NoError(t, err)
-	srvPort, err := strconv.ParseInt(srvURL.Port(), 10, 32)
-	require.NoError(t, err)
-
-	resp, err := srv.Client().Get(srv.URL)
-	require.NoError(t, err)
-	require.NoError(t, resp.Body.Close())
-
-	req := <-got
-	peer, peerPort := splitHostPort(req.RemoteAddr)
-
-	const user = "alice"
-	req.SetBasicAuth(user, "pswrd")
-
-	const clientIP = "127.0.0.5"
-	req.Header.Add("X-Forwarded-For", clientIP)
-
-	srvReq := testServerReq{
-		hostname:   srvURL.Hostname(),
-		serverPort: int(srvPort),
-		peerAddr:   peer,
-		peerPort:   peerPort,
-		clientIP:   clientIP,
-	}
-
-	assert.ElementsMatch(t, want(srvReq), serv.RequestTraceAttrs("", req))
-}
-
-func TestHTTPClientStatus(t *testing.T) {
-	tests := []struct {
-		code int
-		stat codes.Code
-		msg  bool
-	}{
-		{0, codes.Error, true},
-		{http.StatusContinue, codes.Unset, false},
-		{http.StatusSwitchingProtocols, codes.Unset, false},
-		{http.StatusProcessing, codes.Unset, false},
-		{http.StatusEarlyHints, codes.Unset, false},
-		{http.StatusOK, codes.Unset, false},
-		{http.StatusCreated, codes.Unset, false},
-		{http.StatusAccepted, codes.Unset, false},
-		{http.StatusNonAuthoritativeInfo, codes.Unset, false},
-		{http.StatusNoContent, codes.Unset, false},
-		{http.StatusResetContent, codes.Unset, false},
-		{http.StatusPartialContent, codes.Unset, false},
-		{http.StatusMultiStatus, codes.Unset, false},
-		{http.StatusAlreadyReported, codes.Unset, false},
-		{http.StatusIMUsed, codes.Unset, false},
-		{http.StatusMultipleChoices, codes.Unset, false},
-		{http.StatusMovedPermanently, codes.Unset, false},
-		{http.StatusFound, codes.Unset, false},
-		{http.StatusSeeOther, codes.Unset, false},
-		{http.StatusNotModified, codes.Unset, false},
-		{http.StatusUseProxy, codes.Unset, false},
-		{306, codes.Unset, false},
-		{http.StatusTemporaryRedirect, codes.Unset, false},
-		{http.StatusPermanentRedirect, codes.Unset, false},
-		{http.StatusBadRequest, codes.Error, false},
-		{http.StatusUnauthorized, codes.Error, false},
-		{http.StatusPaymentRequired, codes.Error, false},
-		{http.StatusForbidden, codes.Error, false},
-		{http.StatusNotFound, codes.Error, false},
-		{http.StatusMethodNotAllowed, codes.Error, false},
-		{http.StatusNotAcceptable, codes.Error, false},
-		{http.StatusProxyAuthRequired, codes.Error, false},
-		{http.StatusRequestTimeout, codes.Error, false},
-		{http.StatusConflict, codes.Error, false},
-		{http.StatusGone, codes.Error, false},
-		{http.StatusLengthRequired, codes.Error, false},
-		{http.StatusPreconditionFailed, codes.Error, false},
-		{http.StatusRequestEntityTooLarge, codes.Error, false},
-		{http.StatusRequestURITooLong, codes.Error, false},
-		{http.StatusUnsupportedMediaType, codes.Error, false},
-		{http.StatusRequestedRangeNotSatisfiable, codes.Error, false},
-		{http.StatusExpectationFailed, codes.Error, false},
-		{http.StatusTeapot, codes.Error, false},
-		{http.StatusMisdirectedRequest, codes.Error, false},
-		{http.StatusUnprocessableEntity, codes.Error, false},
-		{http.StatusLocked, codes.Error, false},
-		{http.StatusFailedDependency, codes.Error, false},
-		{http.StatusTooEarly, codes.Error, false},
-		{http.StatusUpgradeRequired, codes.Error, false},
-		{http.StatusPreconditionRequired, codes.Error, false},
-		{http.StatusTooManyRequests, codes.Error, false},
-		{http.StatusRequestHeaderFieldsTooLarge, codes.Error, false},
-		{http.StatusUnavailableForLegalReasons, codes.Error, false},
-		{499, codes.Error, false},
-		{http.StatusInternalServerError, codes.Error, false},
-		{http.StatusNotImplemented, codes.Error, false},
-		{http.StatusBadGateway, codes.Error, false},
-		{http.StatusServiceUnavailable, codes.Error, false},
-		{http.StatusGatewayTimeout, codes.Error, false},
-		{http.StatusHTTPVersionNotSupported, codes.Error, false},
-		{http.StatusVariantAlsoNegotiates, codes.Error, false},
-		{http.StatusInsufficientStorage, codes.Error, false},
-		{http.StatusLoopDetected, codes.Error, false},
-		{http.StatusNotExtended, codes.Error, false},
-		{http.StatusNetworkAuthenticationRequired, codes.Error, false},
-		{600, codes.Error, true},
-	}
-
-	for _, test := range tests {
-		t.Run(strconv.Itoa(test.code), func(t *testing.T) {
-			c, msg := HTTPClient{}.Status(test.code)
-			assert.Equal(t, test.stat, c)
-			if test.msg && msg == "" {
-				t.Errorf("expected non-empty message for %d", test.code)
-			} else if !test.msg && msg != "" {
-				t.Errorf("expected empty message for %d, got: %s", test.code, msg)
-			}
-		})
-	}
-}
